backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiohttp==3.9.1
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
packaging==23.2
semver==3.0.0
pyyaml==6.0.1
jinja2==3.1.2
cryptography==41.0.7
requests==2.31.0
jsonschema==4.20.0
openapi-spec-validator==0.7.1
psutil==5.9.6
celery==5.3.4
flower==2.0.1
python-dotenv==1.0.0
email-validator==2.1.0
slowapi==0.1.9
reportlab==4.0.7
pandas==2.1.4
numpy==1.26.2
python-magic==0.4.27
docker==6.1.3
pyyaml-include==1.3.2
croniter==2.0.1

